+++ packages/react/test-app/Pages/ClientSideVisit/Page2.tsx
@@ -1,0 +1,3 @@
+export default ({ baz }: { baz: string }) => {
+  return <div>{baz}</div>
+}