+++ packages/svelte/test-app/Pages/ClientSideVisit/Page1.svelte
@@ -1,8 +1,14 @@
-<script>
+<script lang="ts">
   import { router } from '@inertiajs/svelte'
+  import type { Page } from '@inertiajs/core'
 
-  export let foo;
-  export let bar;
+  interface PageProps {
+    foo: string
+    bar: string
+  }
+
+  export let foo: string;
+  export let bar: string;
 
   let errors = 0;
   let finished = 0;
@@ -11,7 +17,7 @@
   const bagErrors = () => {
     router.replace({
       preserveState: true,
-      props: (props) => ({ ...props, errors: { bag: { foo: 'bar' } } }),
+      props: (props: Page['props']) => ({ ...props, errors: { bag: { foo: 'bar' } } }),
       errorBag: 'bag',
       onError: (err) => {
         errors = Object.keys(err).length
@@ -24,7 +30,7 @@
   const defaultErrors = () => {
     router.replace({
       preserveState: true,
-      props: (props) => ({ ...props, errors: { foo: 'bar', baz: 'qux' } }),
+      props: (props: PageProps) => ({ ...props, errors: { foo: 'bar', baz: 'qux' } }),
       onError: (err) => {
         errors = Object.keys(err).length
       },